+++ aioesphomeapi/_frame_helper/noise_encryption.pxd
